+++ packages/vchart-types/types/compile/compiler.d.ts
@@ -1,16 +1,16 @@
 import type { InteractionSpec, IView } from '@visactor/vgrammar-core';
-import type { CompilerListenerParameters, CompilerModel, IGrammarItem, IRenderContainer, IRenderOption } from './interface';
+import type { CompilerListenerParameters, ICompiler, ICompilerModel, IGrammarItem, IRenderContainer, IRenderOption } from './interface';
 import type { IBoundsLike } from '@visactor/vutils';
 import type { EventSourceType } from '../event/interface';
 import type { IChart } from '../chart/interface';
-import type { IColor, Stage } from '@visactor/vrender-core';
+import type { IColor, IStage } from '@visactor/vrender-core';
 import type { IMorphConfig } from '../animation/spec';
 import type { IVChart } from '../core/interface';
 type EventListener = {
     type: string;
     callback: (...args: any[]) => void;
 };
-export declare class Compiler {
+export declare class Compiler implements ICompiler {
     protected _view: IView;
     getVGrammarView(): IView;
     protected _viewListeners: Map<(...args: any[]) => any, EventListener>;
@@ -23,21 +23,21 @@ export declare class Compiler {
     protected _container: IRenderContainer;
     protected _option: IRenderOption;
     private _released;
-    protected _model: CompilerModel;
+    protected _model: ICompilerModel;
     protected _interactions: (InteractionSpec & {
         seriesId?: number;
         regionId?: number;
     })[];
-    getModel(): CompilerModel;
+    getModel(): ICompilerModel;
     private _compileChart;
     constructor(container: IRenderContainer, option: IRenderOption);
     getRenderer(): import("@visactor/vgrammar-core").IRenderer;
     getCanvas(): HTMLCanvasElement | undefined;
-    getStage(): Stage | undefined;
+    getStage(): IStage | undefined;
     initView(): void;
-    handleStageRender: () => void;
+    protected handleStageRender: () => void;
     private _setCanvasStyle;
-    compileInteractions(): void;
+    protected compileInteractions(): void;
     compile(ctx: {
         chart: IChart;
         vChart: IVChart;