+++ packages/vchart-types/types/component/legend/interface.d.ts
@@ -2,13 +2,16 @@ import type { LegendTitle } from '@visactor/vrender-components';
 import type { Datum, IOrientType, IPadding, IRectMarkSpec, ISymbolMarkSpec, ITextMarkSpec, StringOrNumber } from '../../typings';
 import type { IComponent } from '../interface';
 import type { IComponentSpec } from '../base/interface';
-import type { IDiscreteLegendSpec } from './discrete';
-import type { IColorLegendSpec, ISizeLegendSpec } from './continuous';
+import type { IDiscreteLegendSpec } from './discrete/interface';
+import type { IColorLegendSpec, ISizeLegendSpec } from './continuous/interface';
 export type ILegend = IComponent & {
     getLegendData: () => Datum[];
     getSelectedData: () => StringOrNumber[];
     setSelectedData: (d: StringOrNumber[]) => void;
 };
+export interface IDiscreteLegend extends ILegend {
+    getLegendDefaultData: (originalData?: boolean) => StringOrNumber[];
+}
 export type NoVisibleMarkStyle<T> = Omit<T, 'visible'>;
 export type ITitle = {
     textStyle?: NoVisibleMarkStyle<ITextMarkSpec>;