+++ packages/vchart-types/types/mark/rule.d.ts
@@ -1,8 +1,7 @@
 import type { IRuleMarkSpec } from '../typings/visual';
 import { BaseMark } from './base/base-mark';
-import type { IMarkRaw, IMarkStyle } from './interface';
+import type { IMarkStyle, IRuleMark } from './interface';
 import { MarkTypeEnum } from './interface/type';
-export type IRuleMark = IMarkRaw<IRuleMarkSpec>;
 export declare class RuleMark extends BaseMark<IRuleMarkSpec> implements IRuleMark {
     static readonly type = MarkTypeEnum.rule;
     readonly type = MarkTypeEnum.rule;