+++ packages/vchart-types/types/series/circle-packing/animation.d.ts
@@ -1,5 +1,4 @@
 import type { IAnimationTypeConfig } from '@visactor/vgrammar-core';
-export type CirclePackingMark = 'leaf' | 'nonLeaf' | 'label' | 'nonLeafLabel';
-export type CirclePackingAppearPreset = 'growIn' | 'fadeIn';
+import type { CirclePackingAppearPreset } from './interface';
 export declare const circlePackingPresetAnimation: (preset: CirclePackingAppearPreset) => IAnimationTypeConfig;
 export declare const registerCirclePackingAnimation: () => void;