+++ packages/vchart-types/types/series/gauge/interface.d.ts
@@ -1,8 +1,7 @@
 import type { IMarkSpec, IMarkTheme } from '../../typings/spec';
 import type { IArcMarkSpec, IPathMarkSpec, IRectMarkSpec } from '../../typings/visual';
-import type { IProgressLikeSeriesSpec, IProgressLikeSeriesTheme } from '../polar/progress-like/interface';
+import type { IProgressLikeSeriesSpec, IProgressLikeSeriesTheme, ProgressLikeAppearPreset } from '../polar/progress-like/interface';
 import type { IAnimationSpec } from '../../animation/spec';
-import type { ProgressLikeAppearPreset } from '../polar/progress-like';
 import type { SeriesMarkNameEnum } from '../interface/type';
 import type { ILabelSpec } from '../../component/label/interface';
 export type GaugePointerMarks = 'pointer' | 'pin' | 'pinBackground';