+++ packages/vchart-types/types/series/scatter/animation.d.ts
@@ -1,8 +1,4 @@
 import type { IAnimationTypeConfig } from '@visactor/vgrammar-core';
-export type ScatterMarks = 'point' | 'label';
-export type ScatterAppearPreset = 'scaleIn' | 'fadeIn';
-export interface IScatterAnimationParams {
-    [key: string]: object;
-}
+import type { IScatterAnimationParams, ScatterAppearPreset } from './interface';
 export declare const scatterPresetAnimation: (_params: IScatterAnimationParams, preset: ScatterAppearPreset) => IAnimationTypeConfig;
 export declare const registerScatterAnimation: () => void;