+++ packages/vchart-types/types/series/treemap/animation.d.ts
@@ -1,5 +1,4 @@
 import type { IAnimationTypeConfig } from '@visactor/vgrammar-core';
-export type TreemapMark = 'leaf' | 'nonLeaf';
-export type TreemapAppearPreset = 'growIn' | 'fadeIn';
+import type { TreemapAppearPreset } from './interface';
 export declare const treemapPresetAnimation: (preset: TreemapAppearPreset) => IAnimationTypeConfig;
 export declare const registerTreemapAnimation: () => void;